requirements.txt
nba_api
requests
libsql-experimental==0.0.41
python-dotenv